feat: reuse lote summary tabs component
