+++ my-project/src/app/app/lotes/page.tsx
@@ -4,9 +4,8 @@
 import { useContext, useState, useEffect } from "react";
 import { AuthenticationContext } from "@/app/context/AuthContext";
 import { LoteSelector, Lote } from "@/components/app/lotes/loteselector";
-import { SummaryLote } from "@/components/app/lotes/summarylote";
+import { LoteDataTabs } from "@/components/app/lotes/lotedatatabs";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
 export default function Dashboard() {
   const { data, loading: authLoading } = useContext(AuthenticationContext);
@@ -104,7 +103,6 @@ export default function Dashboard() {
     );
   }
 
-  const loteId = selectedLote?.id ?? "";
 
   return (
     <div className="w-full max-w-6xl mx-auto p-6">
@@ -126,47 +124,7 @@ export default function Dashboard() {
             />
           </CardContent>
         </Card>
-
-        <Tabs defaultValue="resumen">
-          <TabsList className="grid grid-cols-3 mb-4">
-            <TabsTrigger value="resumen">Resumen</TabsTrigger>
-            <TabsTrigger value="datos">Datos</TabsTrigger>
-            <TabsTrigger value="graficos">Gráficos</TabsTrigger>
-          </TabsList>
-
-          <TabsContent value="resumen">
-            {selectedLote ? (
-              <SummaryLote loteId={loteId} />
-            ) : (
-              <div className="text-center text-gray-500">
-                Selecciona un lote para ver su resumen.
-              </div>
-            )}
-          </TabsContent>
-
-          <TabsContent value="datos">
-            {/* Aquí iría tu contenido de “Datos” */}
-            <Card>
-              <CardHeader>
-                <CardTitle>Datos del Lote</CardTitle>
-              </CardHeader>
-              <CardContent>
-                <p>Implementa aquí tu vista de datos.</p>
-              </CardContent>
-            </Card>
-          </TabsContent>
-
-          <TabsContent value="graficos">
-            <Card>
-              <CardHeader>
-                <CardTitle>Gráficos del Lote</CardTitle>
-              </CardHeader>
-              <CardContent>
-                <p>Aquí se generarán los gráficos de conteos del lote.</p>
-              </CardContent>
-            </Card>
-          </TabsContent>
-        </Tabs>
+        <LoteDataTabs empresaId={data.empresaId} lote={selectedLote} />
       </div>
     </div>
   );
